seq_file on windows #51
I don't think this will compile on linux

+++ libs/seq_file/seq_fasta.h
@@ -28,7 +28,8 @@ char seq_next_read_fasta(SeqFile *sf);
 char seq_read_base_fasta(SeqFile *sf, char *c);
 char seq_read_all_bases_fasta(SeqFile *sf, StrBuf *sbuf);
 
-size_t seq_file_write_name_fasta(SeqFile *sf, const char *name);
+unsigned long seq_file_write_name_fasta(SeqFile *sf, const char *name);
+//size_t seq_file_write_name_fasta(SeqFile *sf, const char *name);
 size_t seq_file_write_seq_fasta(SeqFile *sf, const char *seq, size_t str_len);
 
 size_t seq_file_close_write_fasta(SeqFile *sf);
